requirements.txt


google-api-python-client<3.0.0 # For firebase database
gspread<6.0.0 # For google sheet management# TODO breaking change in 6.0.0
gcsa<3.0.0 # For google calendar management
python-dotenv
google-cloud-firestore<3.0.0
